+++ src/main/kotlin/lotto/model/LottoGenerator.kt
@@ -1,12 +1,13 @@
 package lotto.model
+
 import lotto.util.Constant
 
-class LottoGenerator(){
+class LottoGenerator {
     fun generateLotto(): Set<IntRange> {
         return lottos.shuffled().take(Constant.LOTTO_LEN).toSet()
     }
 
-    companion object{
+    companion object {
         private val lottos = listOf(Constant.LOTTO_NUM_RANGE)
     }
 }
