+++ src/main/kotlin/lotto/model/Lottos.kt
@@ -2,6 +2,4 @@ package lotto.model
 
 class Lottos(
     private val lotto: List<Lotto>,
-) {
-
-}
+)
